F16 related code optimization + training vehicle added
