+++ addons/orbis_airforce_config/scripts/fnc_F16AerobaticInit.sqf
@@ -1,0 +1,11 @@
+params ["_vehicle"];
+
+_vehicle animate ["hp2_hide", 1];
+_vehicle animate ["hp3_hide", 1];
+_vehicle animate ["hp7_hide", 1];
+_vehicle animate ["hp8_hide", 1];
+
+_vehicle addEventHandler ["Killed", {[_this select 0, "off"] spawn orbis_airforce_fnc_toggleSmoke;}];
+
+[_vehicle] spawn orbis_airforce_fnc_F16AerobaticLoop;
+[_vehicle] spawn orbis_airforce_fnc_F16AerobaticPostInit;
