--- addons/orbis_airforce_config/scripts/fnc_aerobaticF16Loop.sqf
@@ -1,6 +1,0 @@
-private _vehicle = _this select 0;
-
-[_vehicle,
-["FIR_M61A2", ["FIR_M61A2_511rnd_M", "FIR_240rnd_CMFlare_Chaff_Magazine"]],
-["", ["FIR_240rnd_CMFlare_Chaff_Magazine", "FIR_240rnd_CMFlare_Chaff_Magazine", "FIR_240rnd_CMFlare_Chaff_Magazine", "FIR_240rnd_CMFlare_Chaff_Magazine"]]
-] spawn orbis_airforce_fnc_autoSwapWeapon;
